template refactor
