+++ .github/workflows/chromatic.yml
@@ -2,6 +2,7 @@ name: "Chromatic"
 
 on:
   push:
+    types: [review_requested, ready_for_review]
     paths:
       - "sandpack-react/src/**"
