+++ sandpack-react/src/presets/Sandpack.tsx
@@ -6,71 +6,9 @@ import type { CodeEditorProps } from "../components/CodeEditor";
 import { SandpackCodeEditor } from "../components/CodeEditor";
 import type { PreviewProps } from "../components/Preview";
 import { SandpackPreview } from "../components/Preview";
-import type { SandpackProviderProps } from "../contexts/sandpackContext";
 import { SandpackProvider } from "../contexts/sandpackContext";
-import type {
-  SandpackCodeOptions,
-  FileResolver,
-  SandpackFiles,
-  SandpackInitMode,
-  SandpackPredefinedTemplate,
-  SandpackSetup,
-  SandpackThemeProp,
-} from "../types";
-
-export interface SandpackProps {
-  files?: SandpackFiles;
-  template?: SandpackPredefinedTemplate;
-  customSetup?: SandpackSetup;
-
-  theme?: SandpackThemeProp;
-  sandboxId?: string;
-
-  options?: {
-    openPaths?: string[];
-    activePath?: string;
-
-    editorWidthPercentage?: number;
-    editorHeight?: React.CSSProperties["height"];
-    classes?: Record<string, string>;
-
-    showNavigator?: boolean;
-    showLineNumbers?: boolean;
-    showInlineErrors?: boolean;
-    showTabs?: boolean;
-    closableTabs?: boolean;
-    wrapContent?: boolean;
-
-    /**
-     * This provides a way to control how some components are going to
-     * be initialized on the page. The CodeEditor and the Preview components
-     * are quite expensive and might overload the memory usage, so this gives
-     * a certain control of when to initialize them.
-     */
-    initMode?: SandpackInitMode;
-
-    bundlerURL?: string;
-    startRoute?: string;
-    skipEval?: boolean;
-    fileResolver?: FileResolver;
-    externalResources?: string[];
-
-    autorun?: boolean;
-    recompileMode?: "immediate" | "delayed";
-    recompileDelay?: number;
-    codeEditor?: SandpackCodeOptions;
-
-    /**
-     * This disables editing of content by the user in all files.
-     */
-    readOnly?: boolean;
-    /**
-     * Controls the visibility of Read-only label, which will only
-     * appears when `readOnly` is `true`
-     */
-    showReadOnly?: boolean;
-  };
-}
+import type { SandpackProviderProps } from "../types";
+import type { SandpackProps } from "../types";
 
 /**
  * @category Presets
@@ -128,7 +66,6 @@ export const Sandpack: React.FC<SandpackProps> = (props) => {
   return (
     <SandpackProvider
       customSetup={userInputSetup}
-      sandboxId={props.sandboxId}
       template={props.template}
       {...providerOptions}
     >
